Temporarily disable checks for the latest Buildkit version
Currently, our CI job fails for the latest Buildkit version when
building ARM images. The reason is that an extra field is added in the
config blob of the produced image, when BuildKit v0.20.0 is used.

Temporarily disable checking against the latest BuildKit version, until
the underlying issue is resolved.

Refs #5774